backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
aiosqlite>=0.19,<1.0
pydantic==2.5.3
openai
python-dotenv==1.0.0
transformers==4.36.2
numpy==1.26.3
soundfile==0.12.1 
openai-whisper
pydantic-settings
pydub
backoff
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
python-docx>=1.1.0
openpyxl>=3.1.2
